Populate core modules
