+++ requirements.txt
@@ -1,0 +1,1 @@
+click>=8.0
