+++ weather/__init__.py
@@ -1,0 +1,7 @@
+"""Top-level package for the weather utility."""
+
+from __future__ import annotations
+
+from .core import WeatherError
+
+__all__ = ["WeatherError"]
