+++ weather/commands/__init__.py
@@ -1,0 +1,5 @@
+"""Additional CLI commands for the weather utility."""
+
+# This package can hold click command functions in the future.
+
+__all__: list[str] = []
