+++ weather/core/__init__.py
@@ -1,0 +1,10 @@
+"""Core functionality for the weather utility."""
+
+from __future__ import annotations
+
+
+class WeatherError(Exception):
+    """Custom exception for weather errors."""
+
+
+__all__ = ["WeatherError"]
